Add untested transpiler to fix bug with alternate recipes
selecting a meal with multiple alternate recipes will automatically prioritize the first recipe, and will take and refund ingredients for the first recipe only, but the game can thing that you are allowed to craft meals even if you don't have the ingredients required for the first recipe, only the second.

this patch will take ingredients from the first *available* recipe to craft... just don't try refunding it, or try running this because it probably breaks because i'm terrible at transpilers.